Extract state inspection to state.go
